requirements.txt
Flask==2.3.2
pymongo==4.4.0
python-dotenv==1.0.0
qrcode==7.4.2
Pillow==10.0.0
bcrypt==4.0.1
